API: flattened future streams should not emit if future already resolved
